# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Request:
    properties:
      end_date:
        type: string
      formation_date:
        type: string
      moderator:
        $ref: '#/definitions/ds.User'
      moderatorID:
        type: integer
      request_id:
        type: integer
      start_date:
        type: string
      status:
        type: string
      tasks:
        items:
          $ref: '#/definitions/ds.Task'
        type: array
      user:
        $ref: '#/definitions/ds.User'
      userID:
        type: integer
    type: object
  ds.Status:
    properties:
      status:
        type: string
    type: object
  ds.Task:
    properties:
      description:
        type: string
      image:
        type: string
      minidescription:
        type: string
      name:
        type: string
      status:
        type: string
      subject:
        type: string
      task_id:
        type: integer
    type: object
  ds.User:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        description: Role         role.Role `sql:"type:string;"`
        type: string
      user_id:
        type: integer
    type: object
host: 127.0.0.1
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: BITOP
  version: "1.0"
paths:
...
  /requests:
    get:
      description: Get all requests
      operationId: get-requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Request'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Get Requests
      tags:
      - Requests
  /requests/{id}/admin/update-status:
    put:
      consumes:
      - application/json
      description: Update request by moderator
      operationId: update-request-status-by-moderator
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: status info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Status'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Update Request Status By Moderator
      tags:
      - Requests
  /requests/{id}/user/update-status:
    put:
      consumes:
      - application/json
      description: Update request status by user
      operationId: update-request-status-by-user
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Update Request Status By User
      tags:
      - Requests
  /requests/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete request by ID
      operationId: delete-request-by-id
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Request'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Request'
      security:
      - ApiKeyAuth: []
      summary: Delete Request by ID
      tags:
      - Requests
...
schemes:
- http
swagger: "2.0"